+++ Changelog.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- `ipcores` and `project` `sim` commands updated to Vivado 2017.4
 - Added `--filter` option to `ipbb dep report`
 - Simplified `ipbb vivado project`: IP syntesis runs are not launched automatically anymore.
 - Added `--jobs` option to `ipbb vivado` `synth` and `impl` commands.